Move methods from utils to Sealevel RPC client
